Fixed test crashes (Codex)
