+++ GuiSubtrans/GuiSubtitleTestCase.py
@@ -1,3 +1,5 @@
+from PySide6.QtCore import QCoreApplication
+
 from GuiSubtrans.ProjectDataModel import ProjectDataModel
 from PySubtrans.Helpers.TestCases import BuildSubtitlesFromLineCounts, SubtitleTestCase
 from PySubtrans.Subtitles import Subtitles
@@ -5,6 +7,15 @@ from PySubtrans.Subtitles import Subtitles
 
 class GuiSubtitleTestCase(SubtitleTestCase):
     """Test case with helpers for GUI-level ProjectDataModel interactions."""
+    _qt_app : QCoreApplication|None = None
+
+    @classmethod
+    def setUpClass(cls) -> None:
+        super().setUpClass()
+        if QCoreApplication.instance() is None:
+            cls._qt_app = QCoreApplication([])
+        else:
+            cls._qt_app = QCoreApplication.instance()
 
     def create_project_datamodel(self, subtitles : Subtitles|None = None) -> ProjectDataModel:
         """Create a ProjectDataModel for the provided subtitles."""
